pyproject.toml
[project]
name = "autogenerateatlassianconfluencewikiagent"
version = "0.1.0"
description = "Auto-generate Atlassian Confluence Wiki documentation using multi-agent approach"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.1.0",
    "anthropic>=0.40.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "gitpython>=3.1.0",
    "requests>=2.31.0",
    "asyncio>=3.4.3",
    "aiofiles>=23.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]
